requirements.txt
gradio
biopython
google-genai==1.7.0